requirements.txt
rasterio
esy-osm-pbf